+++ src/tcerror.h
@@ -107,6 +107,7 @@ DEFINE_TCError(CircularMod    , "circular module import: `%s'"                  
 DEFINE_TCError(Required       , "require `%s' type, but is `%s' type"               );
 DEFINE_TCError(CastOp         , "unsupported cast op: `%s' type -> `%s' type"       );
 DEFINE_TCError(UnmatchParam   , "not match parameter, require size is %d, but is %d");
+DEFINE_TCError(RegexSyntax    , "regex syntax error: `%s'"                          );
 
 #undef DEFINE_TCError